Fix a problem with the '.ignore' functionality
Previously, there was a bug in the function 'filter_repos' which
caused the fn to assume that the repo argument was valid if it was
not the first entry in the list of invalid repos. This is now fixed.
Adding names of repositories to the .ignore file causes the tool to
ignore repositories matching this name.