It works now!
This commit marks the stability of the line-wrap() module. Load it
in your .emacs file to use its functions. line-wrap() truncates all
lines which extend beyond the custom variable line-wrap-column-number,
and line-wrap-test-lines() tests all the lines in the region to
determine if they extend beyond the column number. In addition, there
are a number of custom variables which may be changed to tweak the
behavior of the functions to your liking.