Change paths to be relative to MY_GIT env var
This commit changes all paths to be relative to the $MY_GIT variable,
as opposed to the home directory. This is so that I can set MY_GIT
to be the directory containing my git repositories in my .bash_aliases
on my home and machine, and then simply source the file in this repo.