Remove unnecessary code from arg-parse.c & arg-parse.h, add error.h
This patch removes code from the two mentioned files that was used
primarily to perform specific functions in the program from whence
they came. Since this is not that program, I got rid of it. Also
add error.h, a pretty error reporting interface, which prints errors
in the GCC style.