New save-buffer hook and eol script
This patch adds the update-last-edited-date hook to the key binding
C-x C-s, which updates the "LAST EDITED:" field of my stylistic
banner comments at the beginning of the file on save. Also,
matlab-mode.el has officially become unsupported, so we are adding
a new script, matlab.el, which is whole in its functionality.