Update spice-mode with a few stylistic touches
Comment out assignment of buffer local variables `paragraph-start' and
`paragraph-separate.' This allows functions like forward-paragraph to
function normally, but it may affect other functionality of spice-mode.
Change highlighting in comments and doc strings to use `' as
delimiters, instead of ''. This is to make it follow emacs-lisp-mode.
Change doc-string delimiters to use font-lock-string-face. This was a
personal choice. I just think it looks a little more uniform. Change
spice-model-name-face to use "cyan" as a color, instead of "Red3."
This, on a lot of systems, shared the same color as
font-lock-comment-face, which became confusing on the eyes.