Add filtering for 'b update'
This patch adds the ability to add a '.bignore' file to a
repository. The purpose of this file is to filter out tracking of
TODO statements in certain files. It has the same format as .gitignore
files, meaning it is a newline-separated list of filenames. The filter
is regex-based, so BRE regex statements may be used.