Add `sloc' to .bash_aliases
This handy little function computes the number of sloc (system lines of code)
in this and all subdirectories. It takes as an argument a regular expression
to grep the results of `find .'. If no regular expression is given, the default
value is '\.[ch]' (grepping for C source and header files). Also miscellaneous
updates to other files.