Revert Pregit commit
Come to find out, about two months after I made the commit that
added the Pregit tool, that Git already has hooks built in. Needless
to say, I'm a little red-faced about it. This commit removes all of
the additions made in the Pregit commit, and also adds .html files
to the mode list for nxml-mode.