Fix path error on debian and Dockerfile mode
On debian, I experienced an issue where running emacs with sudo does
not load the .emacs initialization file. The changes to these paths
incorporate a fix I made. In addition, when commenting out the line
that loads the dockerfile-mode.el file, emacs encounters an error
due to trying to load a mode it doesn't know about. Just move this
line up into the same area as the load-file call, so we don't forget
about it when we comment out the load-file call.