Fix and update scripts, add disablesleep-tool
This patch fixes a path error in my emacs init file, created by a
previous commit. In addition, update the etwardy-nfs.sh script to
contain the information necessary to connect to my new server.
Finally, add disablesleep-tool.sh, which sets power management
assertions on OS X to prevent the laptop from going to sleep when the
lid is closed.